Move _asserts.ts and _callbacks.ts into mock.ts
